+++ app/_shared/utils/session.ts
@@ -1,51 +1,9 @@
 import type { User } from "@prisma/client";
 import { redirect } from "next/navigation";
-import type { NextAuthOptions } from "next-auth";
-import NextAuth, { getServerSession } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
 import { cache } from "react";
 
-import { userService } from "@/_shared/service/user";
-
-import { createLogger } from "./logger";
+import { getServerSession } from "./getServerSession";
 import { prisma } from "./prisma";
-
-const logger = createLogger("next-auth");
-
-const authOptions: NextAuthOptions = {
-  callbacks: {
-    async session({ session, token }) {
-      session.user = {
-        id: token.id,
-      };
-      return session;
-    },
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-      }
-      return token;
-    },
-  },
-  logger: {
-    error(code, message) {
-      logger.error(`${code} ${JSON.stringify(message)}`);
-    },
-    warn(code) {
-      logger.warn(code);
-    },
-  },
-  providers: [
-    CredentialsProvider({
-      // authorizeの型エラーを解消するために必要
-      // フォーム生成はしないので値は空にしておく
-      credentials: {},
-      authorize: userService.authorize,
-    }),
-  ],
-};
-
-export const handler = NextAuth(authOptions);
 
 const getSessionUserIdOrNull = cache(async () => {
   const session = await getServerSession();
